+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -24,6 +24,7 @@ using B2B.Transactions.Infrastructure.Configuration.Correlation;
 using B2B.Transactions.Infrastructure.DataAccess;
 using B2B.Transactions.Infrastructure.Messages;
 using B2B.Transactions.Infrastructure.Serialization;
+using B2B.Transactions.Infrastructure.Transactions;
 using B2B.Transactions.OutgoingMessages;
 using B2B.Transactions.Xml.Incoming;
 using B2B.Transactions.Xml.Outgoing;
